+++ src/ai/agents/site-spec-agent.js
@@ -57,6 +57,7 @@ class SiteSpecAgent extends BuilderAgent {
 	}
 
 	additionalInstructions( context ) {
+		console.warn( 'formatting', context );
 		return additionalInstructions.format( context );
 	}
